chore: implement simulation

+++ internal/models/review.go
@@ -8,9 +8,9 @@ type Review struct {
 	CustomerID        string    `json:"customer_id"`
 	RestaurantID      string    `json:"restaurant_id"`
 	DeliveryPartnerID string    `json:"delivery_partner_id"`
-	FoodRating        int       `json:"food_rating"`
-	DeliveryRating    int       `json:"delivery_rating"`
-	OverallRating     int       `json:"overall_rating"`
+	FoodRating        float64   `json:"food_rating"`
+	DeliveryRating    float64   `json:"delivery_rating"`
+	OverallRating     float64   `json:"overall_rating"`
 	Comment           string    `json:"comment"`
 	CreatedAt         time.Time `json:"created_at"`
 	UpdatedAt         time.Time `json:"updated_at"`
